more compiler fixes via Claude

+++ tsk/fs/dcat_lib.c
@@ -106,7 +106,7 @@ tsk_fs_blkcat(TSK_FS_INFO * fs, TSK_FS_BLKCAT_FLAG_ENUM lclflags,
 
         /* Read the block */
         cnt = tsk_fs_read_block(fs, addr + i, buf, fs->block_size);
-        if (cnt != fs->block_size) {
+        if (cnt != (ssize_t) fs->block_size) {
             if (cnt >= 0) {
                 tsk_error_reset();
                 tsk_error_set_errno(TSK_ERR_FS_READ);
